requirements.txt
anthropic>=0.40.0
pymupdf>=1.24.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
